requirements.txt
bcrypt==3.2.0
black==21.12b0
cffi==1.15.0
click==8.0.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.1
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
six==1.16.0
SQLAlchemy==1.4.29
tomli==1.2.3
typing-extensions==4.0.1
Werkzeug==2.0.2
WTForms==3.0.1
